requirements.txt
colorama
termcolor
art
python-dotenv==1.0.1